docker/kafka_apps/1_producers/requirements.txt
kafka-python
confluent-kafka
faker
fastavro
requests
jsonschema